autoencoder/requirements.txt
torch
Pillow
streamlit 
opencv-python
